requirements.txt
moviepy==1.0.3
requests~=2.24.0
numpy==1.21.2
beautifulsoup4==4.10.0
PyQt5==5.15.4
system_hotkey==1.0.3
Flask~=2.0.2
bs4~=0.0.1
Pillow~=8.2.0
piexif~=1.1.3
googletrans~=3.0.0
paddlepaddle-gpu==2.1.3.post110
paddlenlp~=2.1.0